Preserve sample rate quality of wideband mixmonitor recordings.
MixMonitor has the ability to record in any file format Asterisk supports,
but the quality of wideband audio is not preserved.  This is because
regardless of the sample rate the call is being recorded in, the audio
is always downsampled to 8khz and then upsampled to whatever wideband
format it is being written as.  This patch resolves this by requesting
the audio from the audiohook in the signed linear format closest to the
sample rate of the format we are writing.  This fix is only possible for
Asterisk 1.10 because audio hooks in 1.8 are not capable of wideband
audio.

Review: https://reviewboard.asterisk.org/r/1314/