Create core configuration files to replace 'make samples'. This commit is the very start of a new configuration file set that users would install and start with rather than the unweildy and incorrectly used 'make samples'. This configuration set starts with autoload=no and then loading all 'core' modules (those marked with a support leve of 'core' in the MODULEINFO section of the module source code).
The approach taken here is that you have only the required .conf files in
the root directory, which then includes files located in two subdirectories:

* default
* site

The default directory contains only the base configuration that marks the
default options, and should generally be considered read-only.

The site directory contains the files that should override any defaults
the administrator wants to change, along with their custom changes.