Its possible to loose audio on ast_write when the channel is not transcoded correctly. in the case of DAHDI the channel is hungup.
This patch tries to "fix" the problem and make the channel compatiable and warn the user of
this problem.

Please note there is a underlying problem with codec negotion this does not fix the problem
it does try to rectify it and prevent loss of service.

Review: https://reviewboard.asterisk.org/r/1442/

(closes issue ASTERISK-17541)
(closes issue ASTERISK-18063)
(issue ASTERISK-14384)
(issue ASTERISK-17502)
(issue ASTERISK-18325)
(issue ASTERISK-18422)