Initial commit of astdb-as-a-realtime-module
Includes a fix to res_config_sqlite3's deltree function that I should fix in
the main tree. Still have some cleanup to do most likely.

The main idea is to eventually phase out the ast_db API and just use the realtime
API. I would like to see the realtime API beefed up a bit too. It is a bit...clunky.

All astdb tests pass. It is currently about 3 times slower than the non-realtime
implementation due to not being able to use SQL statement binding. That is one
of the improvements I'd like to do to realtime--the ability to register queries
so they can be cached.