Add constraint hints to ast_realtime_require_field
The require_field function was insufficient for the requirements of
creating the astdb table since it is important for the backend to return
an error when trying to insert a duplicate key. Also, without an index
on the key, update performance was horrible

This API change required all callers of ast_realtime_require_field to
add a second SENTINEL value if they aren't passing these requirements,
but this was the only way I could think of doing things that wouldn't
require re-writing the realtime backends to support the new feature.