Simplify libasteriskssl wrapping of OpenSSL functions.
The Unix ABI specifies how the dynamic linker behaves; after some research,
it has become apparent that the ABI will ensure that any symbols defined
in libasteriskssl.so *will* be used instead of the same-named symbols defined
in libssl.so, as long as libasteriskssl.so appears in the list of libraries
specified to the linker *before* libssl.so. This means it is safe for the
main Asterisk binary to be linked to both libraries, and thus libasteriskssl.so
does not need to wrap any libssl functions except for the initialization and
shutdown functions.