Undo previous two commits for voicemail.
The scope of what needs to be done needs to be narrowed dramatically,
and so I'm going to focus solely on improving the API to not use message
indexes. The result is going to be a (hopefully temporary) performance
decrease since finding a message by unique ID will require more disk reads
than previously required.