Change Voicemail APIs to use unique message IDs instead of message indexes.
This is not the most efficient operation since messages have to be retrieved
from their storage medium and then read from the file system to be searched. It
would be more efficient to have an in-memory store of the voicemail's current state.
However, doing so is a *HUGE* undertaking, much larger than anything we would be
comfortable adding to Asterisk 11.

In addition, the voicemail API tests have been updated to pass with these API
changes.